part6/flux-architecture-and-redux: 6.7 anecdotes step 5
