+++ part6/redux-anecdotes/src/App.jsx
@@ -1,8 +1,11 @@
 import { useSelector, useDispatch } from "react-redux"
-import { addVotes, addNewAnecdote } from "./reducers/anecdoteReducer"
+
+import { AnecdoteForm } from "./components/AnecdoteForm"
+
+import { addVotes } from "./reducers/anecdoteReducer"
 
 const App = () => {
-  const anecdotes = useSelector((state) => state)
+  const anecdotes = useSelector((state) => state.sort((a, b) => b.votes - a.votes))
   const dispatch = useDispatch()
 
   const vote = (id) => {
@@ -10,37 +13,21 @@ const App = () => {
     dispatch(addVotes(id))
   }
 
-  const handleSubmitAnecdote = (event) => {
-    event.preventDefault()
-    dispatch(addNewAnecdote(event.target.anecdote.value))
-    event.target.anecdote.value = ""
-  }
-
   return (
     <div>
       <h2>Anecdotes</h2>
 
-      {anecdotes
-        .sort((a, b) => b.votes - a.votes)
-        .map((anecdote) => (
-          <div key={anecdote.id}>
-            <div>{anecdote.content}</div>
-            <div>
-              has {anecdote.votes}
-              <button onClick={() => vote(anecdote.id)}>vote</button>
-            </div>
+      {anecdotes.map((anecdote) => (
+        <div key={anecdote.id}>
+          <div>{anecdote.content}</div>
+          <div>
+            has {anecdote.votes}
+            <button onClick={() => vote(anecdote.id)}>vote</button>
           </div>
-        ))}
-
-      <h2>create new</h2>
-
-      <form onSubmit={handleSubmitAnecdote}>
-        <div>
-          <input name="anecdote" />
         </div>
+      ))}
 
-        <button type="submit">create</button>
-      </form>
+      <AnecdoteForm />
     </div>
   )
 }
